Factor out unauthorized user data provider in tests
