+++ tests/Application/FileSource/AbstractFileTest.php
@@ -5,11 +5,14 @@ declare(strict_types=1);
 namespace App\Tests\Application\FileSource;
 
 use App\Tests\Application\AbstractApplicationTestCase;
+use App\Tests\Application\UnauthorizedUserDataProviderTrait;
 use SmartAssert\TestAuthenticationProviderBundle\ApiKeyProvider;
 use Symfony\Component\Uid\Ulid;
 
 abstract class AbstractFileTest extends AbstractApplicationTestCase
 {
+    use UnauthorizedUserDataProviderTrait;
+
     /**
      * @dataProvider unauthorizedUserDataProvider
      */
@@ -66,24 +69,6 @@ abstract class AbstractFileTest extends AbstractApplicationTestCase
         );
 
         self::assertSame(401, $response->getStatusCode());
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function unauthorizedUserDataProvider(): array
-    {
-        return [
-            'no token' => [
-                'token' => null,
-            ],
-            'empty token' => [
-                'token' => '',
-            ],
-            'non-empty invalid token' => [
-                'token' => md5((string) rand()),
-            ],
-        ];
     }
 
     public function testReadSourceNotFound(): void
