+++ tests/Application/FileSource/AbstractListTest.php
@@ -5,12 +5,14 @@ declare(strict_types=1);
 namespace App\Tests\Application\FileSource;
 
 use App\Tests\Application\AbstractApplicationTestCase;
+use App\Tests\Application\UnauthorizedUserDataProviderTrait;
 use Psr\Http\Message\ResponseInterface;
 use SmartAssert\TestAuthenticationProviderBundle\ApiKeyProvider;
 use Symfony\Component\Uid\Ulid;
 
 abstract class AbstractListTest extends AbstractApplicationTestCase
 {
+    use UnauthorizedUserDataProviderTrait;
     use AssertFileSourceTrait;
 
     /**
@@ -21,24 +23,6 @@ abstract class AbstractListTest extends AbstractApplicationTestCase
         $response = $this->applicationClient->makeFileSourceFilesRequest($token, (string) new Ulid());
 
         self::assertSame(401, $response->getStatusCode());
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function unauthorizedUserDataProvider(): array
-    {
-        return [
-            'no token' => [
-                'token' => null,
-            ],
-            'empty token' => [
-                'token' => '',
-            ],
-            'non-empty invalid token' => [
-                'token' => md5((string) rand()),
-            ],
-        ];
     }
 
     public function testListNotFound(): void
