+++ tests/Application/GitSource/AbstractGetTest.php
@@ -5,11 +5,13 @@ declare(strict_types=1);
 namespace App\Tests\Application\GitSource;
 
 use App\Tests\Application\AbstractApplicationTestCase;
+use App\Tests\Application\UnauthorizedUserDataProviderTrait;
 use SmartAssert\TestAuthenticationProviderBundle\ApiKeyProvider;
 use Symfony\Component\Uid\Ulid;
 
 abstract class AbstractGetTest extends AbstractApplicationTestCase
 {
+    use UnauthorizedUserDataProviderTrait;
     use CreateGitSourceDataProviderTrait;
     use AssertGitSourceTrait;
 
@@ -21,24 +23,6 @@ abstract class AbstractGetTest extends AbstractApplicationTestCase
         $response = $this->applicationClient->makeReadGitSourceRequest($token, (string) new Ulid());
 
         self::assertSame(401, $response->getStatusCode());
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function unauthorizedUserDataProvider(): array
-    {
-        return [
-            'no token' => [
-                'token' => null,
-            ],
-            'empty token' => [
-                'token' => '',
-            ],
-            'non-empty invalid token' => [
-                'token' => md5((string) rand()),
-            ],
-        ];
     }
 
     public function testGetNotFound(): void
