+++ tests/Application/UnauthorizedUserDataProviderTrait.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Tests\Application;
+
+trait UnauthorizedUserDataProviderTrait
+{
+    /**
+     * @return array<mixed>
+     */
+    public function unauthorizedUserDataProvider(): array
+    {
+        return [
+            'no token' => [
+                'token' => null,
+            ],
+            'empty token' => [
+                'token' => '',
+            ],
+            'non-empty invalid token' => [
+                'token' => md5((string) rand()),
+            ],
+        ];
+    }
+}
