+++ tests/Application/User/AbstractCreateUserTest.php
@@ -5,9 +5,12 @@ declare(strict_types=1);
 namespace App\Tests\Application\User;
 
 use App\Tests\Application\AbstractApplicationTestCase;
+use App\Tests\Application\UnauthorizedUserDataProviderTrait;
 
 abstract class AbstractCreateUserTest extends AbstractApplicationTestCase
 {
+    use UnauthorizedUserDataProviderTrait;
+
     /**
      * @dataProvider createBadMethodDataProvider
      */
@@ -53,24 +56,6 @@ abstract class AbstractCreateUserTest extends AbstractApplicationTestCase
         );
 
         self::assertSame(401, $response->getStatusCode());
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function unauthorizedUserDataProvider(): array
-    {
-        return [
-            'no admin token' => [
-                'adminToken' => null,
-            ],
-            'empty admin token' => [
-                'adminToken' => '',
-            ],
-            'invalid admin token' => [
-                'adminToken' => md5((string) rand()),
-            ],
-        ];
     }
 
     public function testCreateUserUserAlreadyExists(): void
