+++ tests/Application/User/AbstractListApiKeysTest.php
@@ -5,10 +5,13 @@ declare(strict_types=1);
 namespace App\Tests\Application\User;
 
 use App\Tests\Application\AbstractApplicationTestCase;
+use App\Tests\Application\UnauthorizedUserDataProviderTrait;
 use SmartAssert\TestAuthenticationProviderBundle\FrontendTokenProvider;
 
 abstract class AbstractListApiKeysTest extends AbstractApplicationTestCase
 {
+    use UnauthorizedUserDataProviderTrait;
+
     /**
      * @dataProvider badMethodDataProvider
      */
@@ -45,24 +48,6 @@ abstract class AbstractListApiKeysTest extends AbstractApplicationTestCase
         $response = $this->applicationClient->makeListUserApiKeysRequest($token);
 
         self::assertSame(401, $response->getStatusCode());
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function unauthorizedUserDataProvider(): array
-    {
-        return [
-            'no token' => [
-                'token' => null,
-            ],
-            'empty token' => [
-                'token' => '',
-            ],
-            'non-empty invalid token' => [
-                'token' => md5((string) rand()),
-            ],
-        ];
     }
 
     public function testListSuccess(): void
