+++ tests/Application/User/AbstractRevokeRefreshTokenTest.php
@@ -5,11 +5,14 @@ declare(strict_types=1);
 namespace App\Tests\Application\User;
 
 use App\Tests\Application\AbstractApplicationTestCase;
+use App\Tests\Application\UnauthorizedUserDataProviderTrait;
 use SmartAssert\TestAuthenticationProviderBundle\FrontendTokenProvider;
 use SmartAssert\TestAuthenticationProviderBundle\UserProvider;
 
 abstract class AbstractRevokeRefreshTokenTest extends AbstractApplicationTestCase
 {
+    use UnauthorizedUserDataProviderTrait;
+
     /**
      * @dataProvider badMethodDataProvider
      */
@@ -50,24 +53,6 @@ abstract class AbstractRevokeRefreshTokenTest extends AbstractApplicationTestCas
         $response = $this->applicationClient->makeRevokeAllRefreshTokensForUserRequest($token, md5((string) rand()));
 
         self::assertSame(401, $response->getStatusCode());
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function unauthorizedUserDataProvider(): array
-    {
-        return [
-            'no token' => [
-                'token' => null,
-            ],
-            'empty token' => [
-                'token' => '',
-            ],
-            'non-empty invalid token' => [
-                'token' => md5((string) rand()),
-            ],
-        ];
     }
 
     public function testRevokeRefreshTokenInvalidRefreshToken(): void
